refactor: verificar si la id no es -1 a la hora de borra curso en HomeScreen

+++ app/src/main/java/com/app/grader/ui/pages/home/HomeViewModel.kt
@@ -56,6 +56,7 @@ class HomeViewModel  @Inject constructor(
     }
 
     fun deleteSelectedCourse(onDeleteAction : () -> Unit = {}) {
+        if (_deleteCourseId.intValue == -1) return
         viewModelScope.launch {
             deleteCourseFromIdUseCase(_deleteCourseId.intValue).collect{ result ->
                 when (result){
